Move MediaFileHandler into its own file
